backend/requirements.txt
Flask
pymongo
dnspython
python-dotenv
Werkzeug
gunicorn
